Create and use SocialLinks component
