+++ src/Pages/Contact/index.js
@@ -4,36 +4,11 @@ import { MdEmail } from "react-icons/md"
 
 import Header from '../../components/Header'
 import Hero from '../../components/Hero'
+import SocialLinks from '../../components/SocialLinks'
 import styles from "./styles"
 
 class Contact extends Component {
   render() {
-    const iconAttributes = {
-      color: '#105da1',
-      size: "1.5em",
-    }
-    const links = [
-      {
-        icon: <FaGithub {...iconAttributes} />,
-        text: "Github",
-        url: "https://github.com/apalmer0",
-      },
-      {
-        icon: <FaLinkedin {...iconAttributes} />,
-        text: "LinkedIn",
-        url: "https://www.linkedin.com/in/apalmer0/",
-      },
-      {
-        icon: <MdEmail {...iconAttributes} />,
-        text: "Email",
-        url: "mailto:andrew@andrewpalmer.co",
-      },
-      {
-        icon: <FaPhone {...iconAttributes} />,
-        text: "Phone",
-        url: "tel:617-690-8147",
-      },
-    ]
     return (
       <div style={styles.container}>
         <Header activeTab="contact" />
@@ -53,14 +28,7 @@ class Contact extends Component {
             </p>
           </div>
           <div style={styles.right}>
-            {links.map(({ icon, text, url }) => (
-              <div style={styles.contact} key={text}>
-                <a href={url} target="_blank">
-                  {icon}
-                  <div style={styles.text}>{text}</div>
-                </a>
-              </div>
-            ))}
+            <SocialLinks />
           </div>
         </div>
       </div>
