+++ src/Pages/Contact/styles.js
@@ -21,12 +21,4 @@ export default {
     display: 'inline-block',
     width: '25%',
   },
-  contact: {
-    cursor: 'pointer',
-  },
-  text: {
-    display: 'inline-block',
-    marginLeft: '15px',
-    verticalAlign: 'super',
-  },
 };
