+++ src/components/SocialLinks/styles.js
@@ -1,0 +1,10 @@
+export default {
+  contact: {
+    cursor: 'pointer',
+  },
+  text: {
+    display: 'inline-block',
+    marginLeft: '15px',
+    verticalAlign: 'super',
+  },
+}
